requirements.txt
keyboard
pyautogui
pillow
mss
opencv-python